+++ app/ante.go
@@ -9,6 +9,7 @@ import (
 	ibcante "github.com/cosmos/ibc-go/v7/modules/core/ante"
 	ibckeeper "github.com/cosmos/ibc-go/v7/modules/core/keeper"
 
+	"github.com/cosmos/cosmos-sdk/client"
 	servertypes "github.com/cosmos/cosmos-sdk/server/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	ante "github.com/cosmos/cosmos-sdk/x/auth/ante"
@@ -22,9 +23,20 @@ import (
 	smartaccountante "github.com/osmosis-labs/osmosis/v24/x/smart-account/ante"
 	smartaccountkeeper "github.com/osmosis-labs/osmosis/v24/x/smart-account/keeper"
 
+	auctionkeeper "github.com/skip-mev/block-sdk/x/auction/keeper"
+
 	txfeeskeeper "github.com/osmosis-labs/osmosis/v24/x/txfees/keeper"
 	txfeestypes "github.com/osmosis-labs/osmosis/v24/x/txfees/types"
+
+	auctionante "github.com/skip-mev/block-sdk/x/auction/ante"
 )
+
+// BlockSDKAnteHandlerParams are the parameters necessary to configure the block-sdk antehandlers
+type BlockSDKAnteHandlerParams struct {
+	mevLane       auctionante.MEVLane
+	auctionKeeper auctionkeeper.Keeper
+	txConfig      client.TxConfig
+}
 
 // Link to default ante handler used by cosmos sdk:
 // https://github.com/cosmos/cosmos-sdk/blob/v0.43.0/x/auth/ante/ante.go#L41
@@ -43,6 +55,7 @@ func NewAnteHandler(
 	sigGasConsumer ante.SignatureVerificationGasConsumer,
 	signModeHandler signing.SignModeHandler,
 	channelKeeper *ibckeeper.Keeper,
+	blockSDKParams BlockSDKAnteHandlerParams,
 ) sdk.AnteHandler {
 	mempoolFeeOptions := txfeestypes.NewMempoolFeeOptions(appOpts)
 	mempoolFeeDecorator := txfeeskeeper.NewMempoolFeeDecorator(*txFeesKeeper, mempoolFeeOptions)
@@ -62,6 +75,12 @@ func NewAnteHandler(
 		ante.NewSigVerificationDecorator(accountKeeper, signModeHandler),
 		ante.NewIncrementSequenceDecorator(accountKeeper),
 		ibcante.NewRedundantRelayDecorator(channelKeeper),
+		// auction module antehandler
+		auctionante.NewAuctionDecorator(
+			blockSDKParams.auctionKeeper,
+			blockSDKParams.txConfig.TxEncoder(),
+			blockSDKParams.mevLane,
+		),
 	)
 
 	// authenticatorVerificationDecorator is the new authenticator flow that's embedded into the circuit breaker ante
@@ -72,6 +91,12 @@ func NewAnteHandler(
 		// is embedded in the authenticator decorator
 		smartaccountante.NewAuthenticatorDecorator(smartAccountKeeper, accountKeeper, signModeHandler, deductFeeDecorator),
 		ante.NewIncrementSequenceDecorator(accountKeeper),
+		// auction module antehandler
+		auctionante.NewAuctionDecorator(
+			blockSDKParams.auctionKeeper,
+			blockSDKParams.txConfig.TxEncoder(),
+			blockSDKParams.mevLane,
+		),
 	)
 
 	return sdk.ChainAnteDecorators(