+++ app/modules.go
@@ -64,6 +64,9 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/upgrade"
 	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
 
+	"github.com/skip-mev/block-sdk/x/auction"
+	auctiontypes "github.com/skip-mev/block-sdk/x/auction/types"
+
 	"github.com/osmosis-labs/osmosis/osmoutils/partialord"
 	smartaccount "github.com/osmosis-labs/osmosis/v24/x/smart-account"
 	smartaccounttypes "github.com/osmosis-labs/osmosis/v24/x/smart-account/types"
@@ -135,6 +138,7 @@ var moduleAccountPermissions = map[string][]string{
 	valsetpreftypes.ModuleName:               {authtypes.Staking},
 	poolmanagertypes.ModuleName:              nil,
 	cosmwasmpooltypes.ModuleName:             nil,
+	auctiontypes.ModuleName:                  nil,
 	smartaccounttypes.ModuleName:             nil,
 }
 
@@ -201,6 +205,7 @@ func appModules(
 		packetforward.NewAppModule(app.PacketForwardKeeper, app.GetSubspace(packetforwardtypes.ModuleName)),
 		cwpoolmodule.NewAppModule(appCodec, *app.CosmwasmPoolKeeper),
 		crisis.NewAppModule(app.CrisisKeeper, skipGenesisInvariants, app.GetSubspace(crisistypes.ModuleName)),
+		auction.NewAppModule(appCodec, *app.AuctionKeeper),
 		smartaccount.NewAppModule(appCodec, *app.SmartAccountKeeper),
 	}
 }
@@ -295,6 +300,7 @@ func OrderInitGenesis(allModuleNames []string) []string {
 		icqtypes.ModuleName,
 		packetforwardtypes.ModuleName,
 		cosmwasmpooltypes.ModuleName,
+		auctiontypes.ModuleName,
 	}
 }