+++ x/concentrated-liquidity/genesis_test.go
@@ -43,6 +43,7 @@ var (
 		NextIncentiveRecordId: 2,
 		NextPositionId:        3,
 		IncentivesAccumulatorPoolIdMigrationThreshold: 3,
+		SpreadFactorPoolIdMigrationThreshold:          4,
 	}
 	testCoins    = sdk.NewDecCoins(cl.HundredFooCoins)
 	testTickInfo = model.TickInfo{
@@ -602,6 +603,11 @@ func (s *KeeperTestSuite) TestInitGenesis() {
 			incentiveMigrationThreshold, err := clKeeper.GetIncentivePoolIDMigrationThreshold(ctx)
 			s.Require().NoError(err)
 			s.Require().Equal(tc.genesis.IncentivesAccumulatorPoolIdMigrationThreshold, incentiveMigrationThreshold)
+
+			// Validate spread factor migration threshold
+			spreadFactorMigrationThreshold, err := clKeeper.GetSpreadFactorPoolIDMigrationThreshold(ctx)
+			s.Require().NoError(err)
+			s.Require().Equal(tc.genesis.SpreadFactorPoolIdMigrationThreshold, spreadFactorMigrationThreshold)
 		})
 	}
 }
@@ -824,6 +830,9 @@ func (s *KeeperTestSuite) TestExportGenesis() {
 
 			// Validate incentive migration threshold
 			s.Require().Equal(tc.genesis.IncentivesAccumulatorPoolIdMigrationThreshold, actualExported.IncentivesAccumulatorPoolIdMigrationThreshold)
+
+			// Validate spread factor migration threshold
+			s.Require().Equal(tc.genesis.SpreadFactorPoolIdMigrationThreshold, actualExported.SpreadFactorPoolIdMigrationThreshold)
 		})
 	}
 }